Add the attribute 'type' to set the package URL type
The main objective is to be able to fetch packages that are just a file
which was not possible.

The value of this attribute is either 'file' or 'tarball'. By default,
the value of this attribute is inferred by using the suffix of the
URL. If the URL ends with 'tar.gz' the type is set to 'tarball',
otherwise it is 'file'.

This attribute is used by the Nix expression to select the appropriate
fetcher ('fetchurl' or 'fetchTarball') and by `nix-prefetch-url` to
set or not the `--unpack` flag.