sources.nix: Use the fetcher instead of creating derivations
The created derivations were not enough well created to be handled by
Hydra as a package [1]. Basically, Hydra recurses on the provided
attribute set to figure out which attribute is a derivation. If an
attribute set contains an attribut `type="derivation"`, it is consider
as a derivation. But Hydra requires also more attributes such as a
`name`, `system`...

Since we are now using fetchers from `nixpkgs`, we can directly use
them since they return a derviation.

[1] I add all Niv derivations as packages in Hydra in order to copy
them to the binary cache since it is no done at evaluation time.