Introduce `add` subcommand per fetcher type
This commit introduces subcommand to the `add` command. These
subcommands permit to specify a fetcher type in order to be able to
specify options per fetcher type. For instance, the `--branch` option is
useless for the `file` fetcher. Thanks to this commit, the `--branch`
option is only displayed in the help of the `niv add github` command.

Some usage examples:

    niv add github stedolan/jq
    niv add github NixOS/nixpkgs-channels -n nixpkgs -b nixos-18.09

    niv add file my-package -v alpha-0.1 -t http://example.com/archive/<version>.zip